create table component #38
