+++ client/src/components/pages/ViewBuildings/index.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
-import { Table } from 'antd';
 import axios from 'axios';
 
 import { Navbar } from 'components/utils';
 import styles from './view.module.css';
+import TableInfo from './Table';
 
 class viewBuildings extends Component {
-  state = { buildingInfo: [], filteredInfo: [] };
+  state = { buildingInfo: [] };
 
   async componentDidMount() {
     try {
@@ -19,53 +19,8 @@ class viewBuildings extends Component {
     }
   }
 
-  handleChange = (pagination, filters) => {
-    this.setState({
-      filteredInfo: filters,
-    });
-  };
-
   render() {
-    let { filteredInfo } = this.state;
     const { buildingInfo } = this.state;
-    filteredInfo = filteredInfo || {};
-    const columns = [
-      {
-        title: 'Address',
-        dataIndex: 'address',
-        key: 'address',
-        filters: [
-          { text: 'Morecambe', value: 'Morecambe' },
-          { text: 'Hastings', value: 'Hastings' },
-        ],
-        filteredValue: filteredInfo.address || null,
-        onFilter: (value, record) => record.address.includes(value),
-      },
-      {
-        title: 'Previous use',
-        dataIndex: 'previousUse',
-        key: 'previousUse',
-        filters: [
-          { text: 'Residential building', value: 'Residential building' },
-          { text: 'Retail building', value: 'Retail building' },
-          { text: 'Office building', value: 'Office building' },
-          { text: 'Community building', value: 'Community building' },
-        ],
-        filteredValue: filteredInfo.previousUse || null,
-        onFilter: (value, record) => record.previousUse.includes(value),
-      },
-      {
-        title: 'Owner',
-        dataIndex: 'owner',
-        key: 'owner',
-      },
-      {
-        title: 'Empty period',
-        dataIndex: 'emptyPeriod',
-        key: 'emptyPeriod',
-      },
-    ];
-
     return (
       <>
         <Navbar />
@@ -81,12 +36,7 @@ class viewBuildings extends Component {
           </div>
           <div className="map" />
           <div className="table">
-            <Table
-              columns={columns}
-              dataSource={buildingInfo}
-              onChange={this.handleChange}
-              rowKey={record => record.id}
-            />
+            <TableInfo buildingInfo={buildingInfo} />
           </div>
         </div>
       </>
